[NFC] Moving ast-utils implementation to header
The functionality here is to add helpers to navigate the AST and most
functions are one-liners which will benefit from being inlined. Also, it
allows us to use template parameters to get arguments in a list that is
not defined in the header itself.
